Update KPT_Malawi_1.1.py
I think there is an issue with Firefinder once again. the Cooking events are too long. I think it is something set up in the pre-conditions. Heathers cross sensor anaylsis with the FUEL will not work, But the PM is too variable. I think it going to need more stringent algorithms to compare.